Proje Maven yapısına dönüştürüldü.
Gson kütüphanesi ile JSON veri okuma desteği eklendi.

config.properties dosyasından veri kaynağı (JSON veya TXT) dinamik olarak seçilebilecek şekilde yapılandırıldı.

Exception handling uygulanarak beklenmedik durumlarda programın çökmesi engellendi.

+++ src/main/java/com/sekiluygulamasi/Dortgen.java
@@ -17,8 +17,8 @@ public class Dortgen extends Sekil{
         return 2 * (en + boy);
     }
     public void ciz(){
-        for (int i= 0; i< boy; i++){
-            for (int j= 0; j< en; j++){
+        for (double i= 0; i< boy; i++){
+            for (double j= 0; j< en; j++){
                 System.out.print(sembol + " ");
 
             }
